Add 'protocol' argument and fix hot Thread.Yield loop
This commit adds a 'protocol' field to the command line arguments to
switch between UDP and TCP forwarding.  It also removes an incredibly
hot loop in the main program that called Thread.Yield as fast as
possible this was causing 100% usage of a CPU core.  It was replaced
with a Sleep, which is probably what we wanted in the first place.

[#149819158](https://www.pivotaltracker.com/story/show/149819158)

Signed-off-by: Lisa Oakley <loakley@pivotal.io>