Properly format syslog APP-NAME
We use the event log source as the app-name.  The app-name is not
allowed to contain spaces or be longer than 48 chars.  This commit
replaces spaces with underscores and truncates names longer than 48
chars.  To preserve the original event log source name we added it to
the message JSON under the "source" field.

Signed-off-by: Charlie Vieth <cviethjr@pivotal.io>